# phase-5/specs/2-advanced-cloud-deployment/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Chatbot Chat API
  version: 1.0.0
  description: |
    REST API for the Todo Chatbot Chat API service providing task CRUD operations,
    tag management, and health checks. All task operations publish events to Kafka
    via Dapr Pub/Sub.
  contact:
    name: Todo Chatbot Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://chat-api.todo.svc.cluster.local:8000
    description: Kubernetes cluster
security:
  - bearerAuth: []
paths:
  /tasks:
    post:
      summary: Create a new task
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: List tasks with filtering, search, and sorting
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - name: tags
          in: query
          description: Comma-separated tag names
          schema:
            type: string
        - name: search
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [priority_asc, priority_desc, due_date_asc, due_date_desc, created_asc, created_desc]
            default: created_desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{id}:
    get:
      summary: Get task by ID
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update task fields
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete task
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/complete:
    post:
      summary: Mark task as completed
      operationId: completeTask
      tags:
        - Tasks
      description: Marks task as completed. For recurring tasks, creates next instance.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/TaskResponse'
                  next_task:
                    $ref: '#/components/schemas/TaskResponse'
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        due_at:
          type: string
          format: date-time
        tag_names:
          type: array
          items:
            type: string
        recurrence_pattern:
          $ref: '#/components/schemas/RecurrencePatternCreate'
        reminder_offset_minutes:
          type: integer
          minimum: 5
          maximum: 10080
    RecurrencePatternCreate:
      type: object
      required:
        - pattern_type
        - end_condition
      properties:
        pattern_type:
          type: string
          enum: [daily, weekly, monthly]
        interval:
          type: integer
          minimum: 1
          maximum: 365
          default: 1
        days_of_week:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
        day_of_month:
          type: integer
          minimum: 1
          maximum: 31
        end_condition:
          type: string
          enum: [never, after_occurrences, by_date]
        occurrence_count:
          type: integer
          minimum: 1
        end_date:
          type: string
          format: date-time
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, completed]
        priority:
          type: string
          enum: [high, medium, low]
        due_at:
          type: string
          format: date-time
          nullable: true
        tag_names:
          type: array
          items:
            type: string
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        due_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        recurrence_pattern_id:
          type: integer
          nullable: true
        parent_task_id:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'